group defs together
