+++ runtime/src/handler/prog_handler.hpp
@@ -4,6 +4,8 @@
 #include <boost/interprocess/containers/vector.hpp>
 #include <boost/interprocess/containers/string.hpp>
 #include <ebpf-vm.h>
+#include "bpftime_shm.hpp"
+
 namespace bpftime
 {
 
@@ -27,42 +29,7 @@ class bpf_prog_handler {
 	 * programs correspond to /a/ specific kernel which is to be
 	 * analyzed, and not /a/ specific kernel /and/ all future ones.
 	 */
-	enum class bpf_prog_type {
-		BPF_PROG_TYPE_UNSPEC,
-		BPF_PROG_TYPE_SOCKET_FILTER,
-		BPF_PROG_TYPE_KPROBE,
-		BPF_PROG_TYPE_SCHED_CLS,
-		BPF_PROG_TYPE_SCHED_ACT,
-		BPF_PROG_TYPE_TRACEPOINT,
-		BPF_PROG_TYPE_XDP,
-		BPF_PROG_TYPE_PERF_EVENT,
-		BPF_PROG_TYPE_CGROUP_SKB,
-		BPF_PROG_TYPE_CGROUP_SOCK,
-		BPF_PROG_TYPE_LWT_IN,
-		BPF_PROG_TYPE_LWT_OUT,
-		BPF_PROG_TYPE_LWT_XMIT,
-		BPF_PROG_TYPE_SOCK_OPS,
-		BPF_PROG_TYPE_SK_SKB,
-		BPF_PROG_TYPE_CGROUP_DEVICE,
-		BPF_PROG_TYPE_SK_MSG,
-		BPF_PROG_TYPE_RAW_TRACEPOINT,
-		BPF_PROG_TYPE_CGROUP_SOCK_ADDR,
-		BPF_PROG_TYPE_LWT_SEG6LOCAL,
-		BPF_PROG_TYPE_LIRC_MODE2,
-		BPF_PROG_TYPE_SK_REUSEPORT,
-		BPF_PROG_TYPE_FLOW_DISSECTOR,
-		BPF_PROG_TYPE_CGROUP_SYSCTL,
-		BPF_PROG_TYPE_RAW_TRACEPOINT_WRITABLE,
-		BPF_PROG_TYPE_CGROUP_SOCKOPT,
-		BPF_PROG_TYPE_TRACING,
-		BPF_PROG_TYPE_STRUCT_OPS,
-		BPF_PROG_TYPE_EXT,
-		BPF_PROG_TYPE_LSM,
-		BPF_PROG_TYPE_SK_LOOKUP,
-		BPF_PROG_TYPE_SYSCALL, /* a program that can execute syscalls */
-		BPF_PROG_TYPE_NETFILTER,
-	};
-	enum bpf_prog_type type;
+	bpf_prog_type type;
 
 	bpf_prog_handler(managed_shared_memory &mem,
 			 const struct ebpf_inst *insn, size_t insn_cnt,
